REST - Delete method is working

+++ src/main/java/com/vbilenko/springmvc/service/UserService.java
@@ -21,5 +21,5 @@ public interface UserService {
 	
 	boolean isUserSSOUnique(Integer id, String sso);
 
-    void deleteById(Long id);
+    void deleteById(int id);
 }
